// src/app/pages/program-administration/models/pgm-rvw-info.model.ts
import { RvwType } from "src/app/pages/program-administration/models/rvw-type.model";
import { PgmInfo } from "src/app/pages/program-administration/models/pgm-info.model";
import { RvwTypeCompliancePlanSubmissionReminderSchedule } from "./rvwtype-complianceplan-reminder-schedule.model";
export interface PgmRvwInfo {
  pgmRvwId: number;
  grantPgmId: number;
  rvwType: string;
  wfCd?: string;
  chkListAppl: string;
  rvwTypeAppl: string;
  rvwYrAppl: string;
  rvwCycle?: number;
  resetDttOc: string;
  inclRvwType: string;
  findPkgAppl: string;
  minCntAppl: string;
  minUnannouAppl: string;
  createDt?: Date;
  createId?: number;
  lastUpdDt?: Date;
  lastUpdId?: number;
  recStat: string;
  confirmEmail?: string;
  findingsReport?: string;
  allowAllReviewsRoles?: string;
  allowAllReviewsUsers?: string;
  allowScheduleAllSubs?: string;
  compliancePlanReport?: string;
  allowPreCorrectiveAction: boolean;
  preFindingsReport?: string;
  generateIcsfile?: string;
  hasSiteLevelFindings?: string;
  rvwDesc?: string;
  grantPgm: PgmInfo;
  // rvwTypeNavigation: RvwType;
  rvwTypeCompliancePlanSubmissionReminderSchedules: RvwTypeCompliancePlanSubmissionReminderSchedule[];
}